fix: Display storybook components correctly
fix: Remove unnecessary default export
