+++ packages/storybook/stories/extension/ShortcutLinks.stories.tsx
@@ -32,58 +32,6 @@ export const ListFeedLayout: Story = {
   args: {
     shouldUseListFeedLayout: true,
   },
-  parameters: {
-    msw: {
-      handlers: [
-        http.get('http://localhost:6006/api/boot?v=undefined', async () => {
-          const defaultAlerts: Alerts = {
-            filter: true,
-            rankLastSeen: new Date(),
-          };
-
-          const defaultSettings: RemoteSettings = {
-            theme: 'bright',
-            openNewTab: false,
-            spaciness: 'roomy',
-            insaneMode: false,
-            showTopSites: true,
-            sidebarExpanded: true,
-            companionExpanded: false,
-            sortingEnabled: false,
-            optOutReadingStreak: true,
-            optOutCompanion: true,
-            autoDismissNotifications: true,
-            customLinks: [
-              'http://custom1.com',
-              'http://custom2.com',
-              'http://custom3.com',
-              'http://custom4.com',
-              'http://custom5.com',
-            ],
-            onboardingChecklistView: ChecklistViewState.Hidden,
-          };
-
-          const defaultBootData: BootCacheData = {
-            alerts: defaultAlerts,
-            user: defaultUser,
-            settings: defaultSettings,
-            squads: [],
-            notifications: { unreadNotificationsCount: 0 },
-            feeds: [],
-          };
-
-          const getBootMock = (bootMock: BootCacheData): Boot => ({
-            ...bootMock,
-            accessToken: { token: '1', expiresIn: '1' },
-            visit: { sessionId: '1', visitId: '1' },
-            feeds: [],
-          });
-
-          return HttpResponse.json(getBootMock(defaultBootData));
-        }),
-      ],
-    },
-  },
 };
 
 export const GridFeedLayout: Story = {
